services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pypdf>=4.3.1
python-multipart==0.0.9
alembic==1.13.3
slowapi==0.1.9
google-auth==2.37.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
anyio==4.6.0
